+++ lib/presentation/screen/journal/journal_screen_action.dart
