+++ lib/presentation/screen/journal/journal_screen_state.dart
